useEffect & Dependencies
In this commit, I have used two dependencies in the useEffect hook: enteredEmail and enteredPassword. This ensures that the useEffect function will only run again if both of these values have changed.

In general, useEffect is a super important hook that helps us deal with code that should be executed in response to something. That something could be the component being loaded, the email address being updated, or anything else that triggers an action as a side effect. This is where useEffect can help us.